requirements.txt
python-dotenv==1.0.1
flask==3.0.2
flask-cors==4.0.0
openai==1.40.0
